pyproject.toml
[project]
name = "week2"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.4",
    "django-crispy-forms>=2.4",
]
